+++ AdvancedCooking/CodePatches.cs
@@ -1,6 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using HarmonyLib;
+﻿using System.Collections.Generic;
+using System.Linq;
+using System.Reflection;
+using System.Text;
 using Microsoft.Xna.Framework;
 using StardewValley;
 using StardewValley.Menus;
@@ -8,137 +9,187 @@ using StardewValley.Inventories;
 
 namespace AdvancedCooking
 {
-	/// <summary>The mod entry point.</summary>
 	public partial class ModEntry
 	{
-		private static void CraftingPage_Prefix(int x, int y, bool cooking, List<IInventory> materialContainers, ref int height)
+		class IClickableMenu_populateClickableComponentList_Patch
 		{
-			isCookingMenu = false;
-			containers = null;
-			ingredientMenu = null;
-			if (cooking)
+			public static void Postfix(IClickableMenu __instance)
 			{
-				height += 172;
-				containers = materialContainers;
-				ingredientMenu = new InventoryMenu(x + IClickableMenu.spaceToClearSideBorder + IClickableMenu.borderWidth, y + IClickableMenu.spaceToClearTopBorder + IClickableMenu.borderWidth + Config.YOffset + 84, false, ingredients, null, 11, 1, 0, 0, true);
-				isCookingMenu = true;
-				cookButton = new ClickableTextureComponent(new Rectangle(x + IClickableMenu.spaceToClearSideBorder + IClickableMenu.borderWidth + 64 * 11, y + IClickableMenu.spaceToClearTopBorder + IClickableMenu.borderWidth + Config.YOffset + 80, 64, 64), Game1.mouseCursors, Game1.getSourceRectForStandardTileSheet(Game1.mouseCursors, 46, -1, -1), 1f, false)
-				{
-					myID = 102,
-					rightNeighborID = -99998,
-					leftNeighborID = 106
-				};
-				int labelWidth = (int)Game1.smallFont.MeasureString(SHelper.Translation.Get("fridge-x")).X;
+				if (!Config.ModEnabled || __instance is not CraftingPage craftingPage || !craftingPage.cooking)
+					return;
 
-				fridgeLeftButton = new ClickableTextureComponent("FridgeLeft", new Rectangle(x + IClickableMenu.spaceToClearSideBorder + IClickableMenu.borderWidth - 8, y + IClickableMenu.spaceToClearTopBorder + IClickableMenu.borderWidth + Config.YOffset - 36, 64, 64), null, "", Game1.mouseCursors, Game1.getSourceRectForStandardTileSheet(Game1.mouseCursors, 44, -1, -1), 1f, false)
-				{
-					myID = 631,
-					upNeighborID = -99998,
-					leftNeighborID = -99998,
-					rightNeighborID = -99998,
-					downNeighborID = -99998
-				};
-				fridgeRightButton = new ClickableTextureComponent("FridgeRight", new Rectangle(x + IClickableMenu.spaceToClearSideBorder + IClickableMenu.borderWidth + 128 + labelWidth, y + IClickableMenu.spaceToClearTopBorder + IClickableMenu.borderWidth + Config.YOffset - 36, 64, 64), null, "", Game1.mouseCursors, Game1.getSourceRectForStandardTileSheet(Game1.mouseCursors, 33, -1, -1), 1f, false)
-				{
-					myID = 632,
-					upNeighborID = -99998,
-					leftNeighborID = -99998,
-					rightNeighborID = -99998,
-					downNeighborID = -99998
-				};
+				AttachClickableComponents(craftingPage);
+				AssignClickableComponentNeighbors(craftingPage);
 			}
 		}
 
-		private static void CraftingPage_receiveLeftClick_Prefix(CraftingPage __instance, ref Item ___heldItem, int x, int y)
+		class CraftingPage_Patch
 		{
-			if (!Config.EnableMod || Game1.activeClickableMenu is not CraftingPage || !AccessTools.FieldRefAccess<CraftingPage, bool>(Game1.activeClickableMenu as CraftingPage, "cooking") || ___heldItem is Tool)
-				return;
-			if (cookButton != null && cookButton.containsPoint(x, y))
+			public static void Prefix(CraftingPage __instance, int x, ref int y, bool cooking, List<IInventory> materialContainers, ref int height)
 			{
-				TryCookRecipe(__instance, ref ___heldItem);
-				return;
+				if (!Config.ModEnabled || !cooking)
+					return;
+
+				Reset();
+				Initialize(__instance, x, ref y, materialContainers, ref height);
 			}
-			if (fridgeLeftButton.containsPoint(x, y))
+
+			public static void Postfix(CraftingPage __instance, bool cooking)
 			{
-				Game1.playSound("pickUpItem");
-				fridgeIndex--;
-				if(fridgeIndex < -1)
-					fridgeIndex = __instance._materialContainers.Count - 1;
+				if (!Config.ModEnabled || !cooking)
+					return;
+
+				AttachClickableComponents(__instance);
+				AssignClickableComponentNeighbors(__instance);
 				UpdateActualInventory(__instance);
-				return;
-			}
-			if (fridgeRightButton.containsPoint(x, y))
-			{
-				Game1.playSound("pickUpItem");
-				fridgeIndex++;
-				if (fridgeIndex >= __instance._materialContainers.Count)
-					fridgeIndex = -1;
-				UpdateActualInventory(__instance);
-				return;
-			}
-			___heldItem = ingredientMenu.leftClick(x, y, ___heldItem, true);
-		}
-
-		private static void Game1_drawDialogueBox_Postfix()
-		{
-			if (!Config.EnableMod || Game1.activeClickableMenu is not CraftingPage || !AccessTools.FieldRefAccess<CraftingPage, bool>(Game1.activeClickableMenu as CraftingPage, "cooking"))
-				return;
-			if(currentCookables == null)
-				UpdateCurrentCookables();
-
-			int xStart = Game1.activeClickableMenu.xPositionOnScreen + IClickableMenu.spaceToClearSideBorder + IClickableMenu.borderWidth;
-			int yStart = Game1.activeClickableMenu.yPositionOnScreen + IClickableMenu.borderWidth + IClickableMenu.spaceToClearTopBorder + Config.YOffset;
-			string whichFridge = fridgeIndex >= 0 ? string.Format(SHelper.Translation.Get("fridge-x"), fridgeIndex+1) : SHelper.Translation.Get("player");
-			Utility.drawTextWithShadow(Game1.spriteBatch, whichFridge, Game1.smallFont, new Vector2(xStart + 88, yStart - 20), Color.Black, 1f, -1f, -1, -1, 1f, 3);
-			AccessTools.Method(typeof(IClickableMenu), "drawHorizontalPartition").Invoke(Game1.activeClickableMenu, new object[] { Game1.spriteBatch, yStart, false, -1, -1, -1 });
-			Utility.drawTextWithShadow(Game1.spriteBatch, SHelper.Translation.Get("ingredients") + (Config.ShowProductInfo && currentCookables.Count > 0 ? " " + string.Format(currentCookables.Count == 1 ? SHelper.Translation.Get("will-cook-1") : SHelper.Translation.Get("will-cook-x"), currentCookables.Count) : ""), Game1.smallFont, new Vector2(xStart, yStart + 46), Color.Black, 1f, -1f, -1, -1, 1f, 3);
-			ingredientMenu.draw(Game1.spriteBatch);
-			cookButton.draw(Game1.spriteBatch);
-			fridgeLeftButton.draw(Game1.spriteBatch);
-			fridgeRightButton.draw(Game1.spriteBatch);
-			if (Config.ShowCookTooltip && cookButton.containsPoint(Game1.getMouseX(), Game1.getMouseY()))
-			{
-				DrawCookButtonTooltip();
 			}
 		}
 
-
-		private static void InventoryMenu_Prefix(ref IList<Item> actualInventory)
+		class CraftingPage_receiveLeftClick_Patch
 		{
-			if (!Config.EnableMod || !isCookingMenu || containers == null || containers.Count == 0 || containers[0] == null)
-				return;
-
-			var list = fridgeIndex < 0 ? Game1.player.Items : containers[Math.Min(containers.Count - 1, fridgeIndex)];
-
-			for (int i = 0; i < Game1.player.maxItems.Value; i++)
+			public static void Prefix(CraftingPage __instance, ref Item ___heldItem, int x, int y)
 			{
-				if (list.Count <= i)
-					list.Add(null);
+				if (!Config.ModEnabled || !__instance.cooking)
+					return;
+
+				if (CanCook && CookButton is not null && CookButton.containsPoint(x, y))
+				{
+					TryCook(__instance, ref ___heldItem);
+					return;
+				}
+				if (__instance._materialContainers is not null && __instance._materialContainers.Count > 0)
+				{
+					if (FridgeLeftButton.containsPoint(x, y))
+					{
+						Game1.playSound("pickUpItem");
+						FridgeIndex--;
+						if (FridgeIndex < -1)
+						{
+							FridgeIndex = __instance._materialContainers.Count - 1;
+						}
+						UpdateActualInventory(__instance);
+						return;
+					}
+					if (FridgeRightButton.containsPoint(x, y))
+					{
+						Game1.playSound("pickUpItem");
+						FridgeIndex++;
+						if (FridgeIndex >= __instance._materialContainers.Count)
+						{
+							FridgeIndex = -1;
+						}
+						UpdateActualInventory(__instance);
+						return;
+					}
+				}
+				if (IngredientMenu.isWithinBounds(x, y) && ___heldItem is not Tool)
+				{
+					___heldItem = IngredientMenu.leftClick(x, y, ___heldItem, true);
+					UpdateCookableRecipes();
+				}
 			}
-			actualInventory = list;
-			isCookingMenu = false;
 		}
 
-		private static void CookingMenu_DrawActualInventory_Postfix()
+		class CraftingPage_receiveRightClick_Patch
 		{
-			if (!Config.EnableMod)
-				return;
-			if (currentCookables == null)
-				UpdateCurrentCookables();
-
-			int xStart = Game1.activeClickableMenu.xPositionOnScreen + IClickableMenu.spaceToClearSideBorder + IClickableMenu.borderWidth;
-			int yStart = Game1.activeClickableMenu.yPositionOnScreen + IClickableMenu.borderWidth + IClickableMenu.spaceToClearTopBorder + Config.YOffset;
-			string whichFridge = fridgeIndex >= 0 ? string.Format(SHelper.Translation.Get("fridge-x"), fridgeIndex + 1) : SHelper.Translation.Get("player");
-			Utility.drawTextWithShadow(Game1.spriteBatch, whichFridge, Game1.smallFont, new Vector2(xStart + 88, yStart - 20), Color.Black, 1f, -1f, -1, -1, 1f, 3);
-			AccessTools.Method(typeof(IClickableMenu), "drawHorizontalPartition").Invoke(Game1.activeClickableMenu, new object[] { Game1.spriteBatch, yStart, false, -1, -1, -1 });
-			Utility.drawTextWithShadow(Game1.spriteBatch, SHelper.Translation.Get("ingredients") + (Config.ShowProductInfo && currentCookables.Count > 0 ? " " + string.Format(currentCookables.Count == 1 ? SHelper.Translation.Get("will-cook-1") : SHelper.Translation.Get("will-cook-x"), currentCookables.Count) : ""), Game1.smallFont, new Vector2(xStart, yStart + 46), Color.Black, 1f, -1f, -1, -1, 1f, 3);
-			ingredientMenu.draw(Game1.spriteBatch);
-			cookButton.draw(Game1.spriteBatch);
-			fridgeLeftButton.draw(Game1.spriteBatch);
-			fridgeRightButton.draw(Game1.spriteBatch);
-			if (Config.ShowCookTooltip && cookButton.containsPoint(Game1.getMouseX(), Game1.getMouseY()))
+			public static void Prefix(CraftingPage __instance, ref Item ___heldItem, int x, int y)
 			{
-				DrawCookButtonTooltip();
+				if (!Config.ModEnabled || !__instance.cooking)
+					return;
+
+				if (CanCook && CookButton is not null && CookButton.containsPoint(x, y))
+				{
+					TryCook(__instance, ref ___heldItem);
+					return;
+				}
+				if (IngredientMenu.isWithinBounds(x, y) && ___heldItem is not Tool)
+				{
+					___heldItem = IngredientMenu.rightClick(x, y, ___heldItem, true);
+					UpdateCookableRecipes();
+				}
+			}
+		}
+
+		class CraftingPage_gameWindowSizeChanged_Patch
+		{
+			public static void Postfix(CraftingPage __instance)
+			{
+				if (!Config.ModEnabled || !__instance.cooking)
+					return;
+
+				if (FridgeLeftButton is not null && FridgeRightButton is not null)
+				{
+					FridgeLeftButton.bounds = new Rectangle(__instance.xPositionOnScreen + IClickableMenu.spaceToClearSideBorder + IClickableMenu.borderWidth, __instance.yPositionOnScreen + IClickableMenu.spaceToClearTopBorder + IClickableMenu.borderWidth + Config.YOffset - 36, 64, 64);
+					FridgeRightButton.bounds = new Rectangle(__instance.xPositionOnScreen + IClickableMenu.spaceToClearSideBorder + IClickableMenu.borderWidth + 318, __instance.yPositionOnScreen + IClickableMenu.spaceToClearTopBorder + IClickableMenu.borderWidth + Config.YOffset - 36, 64, 64);
+				}
+				IngredientMenu.SetPosition(__instance.xPositionOnScreen + IClickableMenu.spaceToClearSideBorder + IClickableMenu.borderWidth, __instance.yPositionOnScreen + IClickableMenu.spaceToClearTopBorder + IClickableMenu.borderWidth + Config.YOffset + 84);
+				CookButton.bounds = new Rectangle(__instance.xPositionOnScreen + IClickableMenu.spaceToClearSideBorder + IClickableMenu.borderWidth + 64 * 11 + 8, __instance.yPositionOnScreen + IClickableMenu.spaceToClearTopBorder + IClickableMenu.borderWidth + Config.YOffset + 80, 64, 64);
+				AssignClickableComponentNeighbors(__instance);
+			}
+		}
+
+		class CraftingPage_emergencyShutDown_Patch
+		{
+			public static void Prefix(CraftingPage __instance)
+			{
+				if (!Config.ModEnabled || !__instance.cooking)
+					return;
+
+				HandleBeforeCleanup(__instance);
+			}
+		}
+
+		class Game1_drawDialogueBox_Patch
+		{
+			public static void Postfix()
+			{
+				if (!Config.ModEnabled || Game1.activeClickableMenu is not CraftingPage craftingPage || !craftingPage.cooking)
+					return;
+
+				int xStart = Game1.activeClickableMenu.xPositionOnScreen + IClickableMenu.spaceToClearSideBorder + IClickableMenu.borderWidth;
+				int yStart = Game1.activeClickableMenu.yPositionOnScreen + IClickableMenu.borderWidth + IClickableMenu.spaceToClearTopBorder + Config.YOffset;
+				string whichFridge = FridgeIndex >= 0 ? string.Format(SHelper.Translation.Get("fridge-x"), FridgeIndex + 1) : SHelper.Translation.Get("player");
+
+				typeof(IClickableMenu).GetMethod("drawHorizontalPartition", BindingFlags.NonPublic | BindingFlags.Instance).Invoke(Game1.activeClickableMenu, new object[] { Game1.spriteBatch, yStart, false, -1, -1, -1 });
+				Utility.drawTextWithShadow(Game1.spriteBatch, SHelper.Translation.Get("ingredients") + (Config.ShowProductInfo && CookableRecipes.Count > 0 ? " " + string.Format(CookableRecipes.Count == 1 ? SHelper.Translation.Get("will-cook-1") : SHelper.Translation.Get("will-cook-x"), CookableRecipes.Count) : ""), Game1.smallFont, new Vector2(xStart, yStart + 46), Color.Black, 1f, -1f, -1, -1, 1f, 3);
+				IngredientMenu.draw(Game1.spriteBatch);
+				CookButton.draw(Game1.spriteBatch, CanCook ? Color.White : (Color.Gray * 0.8f), 0.88f);
+				if (FridgeLeftButton is not null && FridgeRightButton is not null)
+				{
+					FridgeLeftButton.draw(Game1.spriteBatch);
+					FridgeRightButton.draw(Game1.spriteBatch);
+					Utility.drawTextWithShadow(Game1.spriteBatch, whichFridge, Game1.smallFont, new Vector2(xStart + 88, yStart - 20), Color.Black, 1f, -1f, -1, -1, 1f, 3);
+				}
+				if (Config.ShowCookTooltip && CanCook && CookButton.containsPoint(Game1.getMouseX(), Game1.getMouseY()) && CookableRecipes.Count > 0)
+				{
+					StringBuilder tooltip = new();
+					List<(Item, int)> values = CookableRecipes.Values.ToList();
+
+					if (!SHelper.Input.IsDown(Config.CookAllModKey))
+					{
+						tooltip.Append(string.Format(SHelper.Translation.Get("x-of-y"), 1, Config.ShowProductsInTooltip ? values[0].Item1.DisplayName : "???"));
+					}
+					else
+					{
+						for (int i = 0; i < values.Count; i++)
+						{
+							if (i > 0)
+							{
+								tooltip.Append('\n');
+							}
+							if (i < Config.MaxItemsInTooltip)
+							{
+								tooltip.Append(string.Format(SHelper.Translation.Get("x-of-y"), values[i].Item2, Config.ShowProductsInTooltip ? values[i].Item1.DisplayName : "???"));
+							}
+							else
+							{
+								tooltip.Append(string.Format(SHelper.Translation.Get("plus-x-more"), CookableRecipes.Keys.Count - i));
+								break;
+							}
+						}
+					}
+					IClickableMenu.drawHoverText(Game1.spriteBatch, tooltip, Game1.smallFont, 0, 0, -1, SHelper.Translation.Get("cook"));
+				}
 			}
 		}
 	}