+++ AdvancedCooking/release-notes.md
@@ -2,6 +2,23 @@
 
 # Release notes
 
+## 0.1.5-unofficial.2-mouahrara
+Released on 11 November, 2024, for SMAPI 4.1.0 or later.
+* Optimized code for improved performance ✨
+* It is now possible to hold down the right-click on the cook button to continuously cook items ✨
+* Ingredients are now automatically returned to the inventory or fridge when the cooking menu is closed ✨
+* Improved GMCM support ✨
+* Added controller support 🎮
+* Fixed an issue where the number of dishes displayed before bulk cooking did not correspond to the actual number of dishes cooked 🔧
+* Fixed an issue where learning new recipes did not correctly update the recipes layout 🔧
+* Fixed an issue where the mod was not working with the cookout kit 🔧
+* Fixed an issue where the names of the dishes were not displayed before cooking 🔧
+* Fixed an issue where the **Show Products In Tooltip** option was not working correctly 🔧
+* Fixed an issue where the number of different dishes in the tooltip exceeded the value of the **Max Types In Tooltip** option 🔧
+* Fixed an issue where the Advanced Cooking menu was not centered 🔧
+* Fixed an issue where the Advanced Cooking menu was not displayed correctly after resizing the window 🔧
+* Fixed bugs in local multiplayer 🔧
+
 ## 0.1.5-unofficial.1-mouahrara
 Released on 24 March, 2024, for SMAPI 4.0.0 or later.
 * Migrated to 1.6.0 and project cleanup 🚀